Use double again instead of BigDecimal

+++ TP4-MarsTravel/src/main/java/itba/ss/TP4_MarsTravel/Accelerator.java
@@ -1,12 +1,11 @@
 package itba.ss.TP4_MarsTravel;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface Accelerator {
 
-	public BigDecimal getForceX(Particle p, List<Particle> l);
+	public double getForceX(Particle p, List<Particle> l);
 
-	public BigDecimal getForceY(Particle p, List<Particle> l);
+	public double getForceY(Particle p, List<Particle> l);
 
 }
